+++ thoth-api/migrations/0.4.5/up.sql
@@ -1,0 +1,3 @@
+ALTER TABLE work
+    ALTER COLUMN width TYPE double precision,
+    ALTER COLUMN height TYPE double precision;