+++ thoth-api/src/model/mod.rs
@@ -3,12 +3,28 @@ use isbn2::Isbn13;
 use serde::{Deserialize, Serialize};
 use std::fmt;
 use std::str::FromStr;
+use strum::Display;
+use strum::EnumString;
 use thoth_errors::{ThothError, ThothResult};
 #[cfg(feature = "backend")]
 use uuid::Uuid;
 
 pub const DOI_DOMAIN: &str = "https://doi.org/";
 pub const ORCID_DOMAIN: &str = "https://orcid.org/";
+
+#[cfg_attr(
+    feature = "backend",
+    derive(juniper::GraphQLEnum),
+    graphql(description = "Unit of measurement for physical Work dimensions (mm, cm or in)")
+)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, EnumString, Display)]
+#[serde(rename_all = "SCREAMING_SNAKE_CASE")]
+#[strum(serialize_all = "lowercase")]
+pub enum LengthUnit {
+    Mm,
+    Cm,
+    In,
+}
 
 #[cfg_attr(
     feature = "backend",
@@ -47,6 +63,12 @@ pub struct Orcid(String);
 )]
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub struct Timestamp(DateTime<Utc>);
+
+impl Default for LengthUnit {
+    fn default() -> LengthUnit {
+        LengthUnit::Mm
+    }
+}
 
 impl Default for Doi {
     fn default() -> Doi {
@@ -407,6 +429,36 @@ macro_rules! db_insert {
     };
 }
 
+pub trait Convert {
+    fn convert_units_from_to(&self, current_units: &LengthUnit, new_units: &LengthUnit) -> f64;
+}
+
+impl Convert for f64 {
+    fn convert_units_from_to(&self, current_units: &LengthUnit, new_units: &LengthUnit) -> f64 {
+        match (current_units, new_units) {
+            // If current units and new units are the same, no conversion is needed
+            (LengthUnit::Mm, LengthUnit::Mm)
+            | (LengthUnit::Cm, LengthUnit::Cm)
+            | (LengthUnit::In, LengthUnit::In) => *self,
+            // Return cm values rounded to max 1 decimal place (1 cm = 10 mm)
+            (LengthUnit::Mm, LengthUnit::Cm) => self.round() / 10.0,
+            // Return mm values rounded to nearest mm (1 cm = 10 mm)
+            (LengthUnit::Cm, LengthUnit::Mm) => (self * 10.0).round(),
+            // Return inch values rounded to 2 decimal places (1 inch = 25.4 mm)
+            (LengthUnit::Mm, LengthUnit::In) => {
+                let unrounded_inches = self / 25.4;
+                // To round to a non-integer scale, multiply by the appropriate factor,
+                // round to the nearest integer, then divide again by the same factor
+                (unrounded_inches * 100.0).round() / 100.0
+            }
+            // Return mm values rounded to nearest mm (1 inch = 25.4 mm)
+            (LengthUnit::In, LengthUnit::Mm) => (self * 25.4).round(),
+            // We don't currently support conversion between cm and in as it is not required
+            _ => unimplemented!(),
+        }
+    }
+}
+
 #[test]
 fn test_doi_default() {
     let doi: Doi = Default::default();
@@ -599,4 +651,80 @@ fn test_orcid_fromstr() {
     assert!(Orcid::from_str("//orcid.org/0000-0002-1234-5678").is_err());
     assert!(Orcid::from_str("https://orcid-org/0000-0002-1234-5678").is_err());
     assert!(Orcid::from_str("0000-0002-1234-5678https://orcid.org/").is_err());
+}
+
+#[test]
+// Float equality comparison is fine here because the floats
+// have already been rounded by the functions under test
+#[allow(clippy::float_cmp)]
+fn test_convert_units_from_to() {
+    use LengthUnit::*;
+    assert_eq!(123.456.convert_units_from_to(&Mm, &Cm), 12.3);
+    assert_eq!(123.456.convert_units_from_to(&Mm, &In), 4.86);
+    assert_eq!(123.456.convert_units_from_to(&Cm, &Mm), 1235.0);
+    assert_eq!(123.456.convert_units_from_to(&In, &Mm), 3136.0);
+    // Test some standard print sizes
+    assert_eq!(4.25.convert_units_from_to(&In, &Mm), 108.0);
+    assert_eq!(108.0.convert_units_from_to(&Mm, &In), 4.25);
+    assert_eq!(6.0.convert_units_from_to(&In, &Mm), 152.0);
+    assert_eq!(152.0.convert_units_from_to(&Mm, &In), 5.98);
+    assert_eq!(8.5.convert_units_from_to(&In, &Mm), 216.0);
+    assert_eq!(216.0.convert_units_from_to(&Mm, &In), 8.5);
+    // Test that converting and then converting back again
+    // returns a value within a reasonable margin of error
+    assert_eq!(
+        5.06.convert_units_from_to(&In, &Mm)
+            .convert_units_from_to(&Mm, &In),
+        5.08
+    );
+    assert_eq!(
+        6.5.convert_units_from_to(&In, &Mm)
+            .convert_units_from_to(&Mm, &In),
+        6.5
+    );
+    assert_eq!(
+        7.44.convert_units_from_to(&In, &Mm)
+            .convert_units_from_to(&Mm, &In),
+        7.44
+    );
+    assert_eq!(
+        8.27.convert_units_from_to(&In, &Mm)
+            .convert_units_from_to(&Mm, &In),
+        8.27
+    );
+    assert_eq!(
+        9.0.convert_units_from_to(&In, &Mm)
+            .convert_units_from_to(&Mm, &In),
+        9.02
+    );
+    assert_eq!(
+        10.88
+            .convert_units_from_to(&In, &Mm)
+            .convert_units_from_to(&Mm, &In),
+        10.87
+    );
+    assert_eq!(
+        102.0
+            .convert_units_from_to(&Mm, &In)
+            .convert_units_from_to(&In, &Mm),
+        102.0
+    );
+    assert_eq!(
+        120.0
+            .convert_units_from_to(&Mm, &In)
+            .convert_units_from_to(&In, &Mm),
+        120.0
+    );
+    assert_eq!(
+        168.0
+            .convert_units_from_to(&Mm, &In)
+            .convert_units_from_to(&In, &Mm),
+        168.0
+    );
+    assert_eq!(
+        190.0
+            .convert_units_from_to(&Mm, &In)
+            .convert_units_from_to(&In, &Mm),
+        190.0
+    );
 }